refactor: app에 lint --fix 적용
