+++ apps/app/__tests__/App.test.tsx
@@ -4,13 +4,15 @@
 
 import 'react-native';
 import React from 'react';
+import renderer from 'react-test-renderer';
+
+import {it} from '@jest/globals';
+
 import App from '../App';
 
 // Note: import explicitly to use the types shipped with jest.
-import {it} from '@jest/globals';
 
 // Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
   renderer.create(<App />);
