+++ apps/app/index.js
@@ -3,6 +3,7 @@
  */
 
 import { AppRegistry } from 'react-native'
+
 import App from './App'
 import { name as appName } from './app.json'
 import './src/i18n'
