+++ apps/app/metro.config.js
@@ -1,4 +1,5 @@
 const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config')
+
 const path = require('path')
 
 /**
