+++ apps/app/src/screens/BottomTab/Second/index.tsx
@@ -2,7 +2,7 @@ import { Text, View } from 'react-native'
 
 export type SecondProps = undefined
 
-export const Second = () => {
+export function Second() {
   return (
     <View style={{ backgroundColor: 'yellow', width: 100, height: 100, margin: 100 }}>
       <Text>second</Text>
