+++ apps/app/src/screens/BottomTab/index.tsx
@@ -1,6 +1,7 @@
 import { BottomTabNavigationOptions, createBottomTabNavigator } from '@react-navigation/bottom-tabs'
-import TabBar from './TabBar'
+
 import { BOTTOM_TAB_LIST } from './constants'
+import TabBar from './TabBar'
 import { BOTTOM_TAP_ROUTE_TYPE } from './type'
 
 const BottomTab = createBottomTabNavigator<BOTTOM_TAP_ROUTE_TYPE>()
@@ -9,7 +10,7 @@ const defaultOptions: BottomTabNavigationOptions = {
   headerShown: false,
 }
 
-const BottomTabNavigation = () => {
+function BottomTabNavigation() {
   return (
     <BottomTab.Navigator screenOptions={defaultOptions} tabBar={(props) => <TabBar {...props} />}>
       {BOTTOM_TAB_LIST.map((item) => (
