+++ apps/app/src/screens/BottomTab/type.ts
@@ -1,5 +1,5 @@
-import { FirstProps } from './First'
-import { SecondProps } from './Second'
+import { FirstProps } from '@/screens/BottomTab/First'
+import { SecondProps } from '@/screens/BottomTab/Second'
 
 export type BOTTOM_TAP_ROUTE_TYPE = {
   First: FirstProps
