+++ apps/app/src/screens/RootNativeStack/type.ts
@@ -1,4 +1,7 @@
 import { NavigatorScreenParams } from '@react-navigation/native'
+
+import { ExampleProps } from '@/screens/RootNativeStack/Example'
+
 import { BOTTOM_TAP_ROUTE_TYPE } from '../BottomTab/type'
 
 export type NATIVE_STACK_LIST_TYPE = {
