+++ apps/app/src/screens/type.ts
@@ -1,3 +1,3 @@
-import { NATIVE_STACK_LIST_TYPE } from './RootNativeStack/type'
+import { NATIVE_STACK_LIST_TYPE } from '@/screens/RootNativeStack/type'
 
 export type AllScreenType = NATIVE_STACK_LIST_TYPE
