+++ apps/app/src/shared/global-component/index.tsx
@@ -1,11 +1,12 @@
 import React, { PropsWithChildren } from 'react'
+
 import { Toast } from './Toast'
 
-const GlobalComponentProvider = ({ children }: PropsWithChildren) => (
-  <>
+function GlobalComponentProvider({ children }: PropsWithChildren) {
+  return <>
     {children}
     <Toast.Portal />
   </>
-)
+}
 
 export default GlobalComponentProvider
