+++ apps/app/src/shared/hooks/useAppRouter.ts
@@ -1,6 +1,7 @@
 import { RouteProp, useNavigation, useRoute } from '@react-navigation/native'
-import { AllScreenType } from '../../screens/type'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
+
+import { AllScreenType } from '@/screens/type'
 
 export const useAppRouter = (_route?: string) => {
   const navigation = useNavigation<NativeStackNavigationProp<AllScreenType>>()
