+++ apps/app/src/shared/storage/Auth/index.ts
@@ -1,15 +1,16 @@
-import MMKVWrapper from '..';
-import {STORAGE_KEY} from '../constants';
+import { MMKVWrapper } from '@/shared/storage'
+import { STORAGE_KEY } from '@/shared/storage/constants'
 
 class AuthStorage extends MMKVWrapper<string> {
   setToken(item: string) {
-    this.setItem(item);
+    this.setItem(item)
   }
+
   getToken() {
-    return this.getItem() ?? '';
+    return this.getItem() ?? ''
   }
 }
 
-const authStorage = new AuthStorage(STORAGE_KEY.auth);
+const authStorage = new AuthStorage(STORAGE_KEY.auth)
 
-export default authStorage;
+export default authStorage
