+++ apps/app/src/shared/storage/Language/index.ts
@@ -1,12 +1,15 @@
-import { LanguageTypes } from '@pinit/shared/src/constants'
-import MMKVWrapper from '..'
-import { STORAGE_KEY } from '../constants'
 import { NativeModules, Platform } from 'react-native'
+
+import { LanguageTypes } from '@pinit/shared/src/constants'
+
+import { MMKVWrapper } from '@/shared/storage'
+import { STORAGE_KEY } from '@/shared/storage/constants'
 
 class LanguageStroage extends MMKVWrapper<LanguageTypes> {
   setLanguage(item: LanguageTypes) {
     this.setItem(item)
   }
+
   getLanguage() {
     return this.getItem() ?? this.getSystemOriginLanguage()
   }
@@ -18,11 +21,11 @@ class LanguageStroage extends MMKVWrapper<LanguageTypes> {
         NativeModules.SettingsManager.settings.AppleLanguages[0]
 
       return lang.toLowerCase().indexOf('ko') >= 0 ? 'ko' : 'en'
-    } else {
-      const lang = NativeModules.I18nManager.localeIdentifier
-
-      return lang.toLowerCase().indexOf('ko') >= 0 ? 'ko' : 'en'
     }
+
+    const lang = NativeModules.I18nManager.localeIdentifier
+
+    return lang.toLowerCase().indexOf('ko') >= 0 ? 'ko' : 'en'
   }
 }
