+++ apps/app/src/shared/trpc/index.ts
@@ -1,7 +1,9 @@
-import { createTRPCProxyClient } from '@trpc/client'
 import { AppRouter } from '@pinit/server/src/trpc/trpc.router'
+
+import { createTRPCProxyClient } from '@trpc/client'
 import { createTRPCReact, httpBatchLink } from '@trpc/react-query'
-import authStorage from '../storage/Auth'
+
+import authStorage from '@/shared/storage/Auth'
 
 export const trpcOption = {
   links: [
